refactor: extracted `update()` function
